+++ packages/core/src/index.ts
@@ -72,4 +72,5 @@ export * from "./sprint.ts";
 export * from "./timeline.ts";
 export * from "./opwords.ts";
 export * from "./canvassort.ts";
+export * from "./canvasswitch.ts";
 export * from "./lens.ts";